refactor(KEN-1236): forced splits rejoin their seam [no-changelog]
The size ratchet held source files to a line cap, and a file that hit
the cap was split where the cap fell. The cap is gone, so a split whose
boundary is not a design boundary is put back together:

- rules_markdown.rs asked the same question as rules_blocks.rs, how far
  a code span reaches; its boundary was "what the old walk modelled",
  and the walk no longer exists.
- audit-fold.ts was one function with one importer; copy-footer.ts was
  two labels cut from copy.ts.
- pr-watch-help.test.sh, queue_wait_args.sh and ci_wait_args.sh were
  cases cut from their script's suite under the frozen test class.

Every other cap-triggered split in the range was judged and left where
it is; the table is on KEN-1236. Cases move whole, so each keeps the
must-fail direction it was written with.
