+++ crates/core/tests/quality/main.rs
@@ -14,6 +14,5 @@ mod reading;
 mod rules;
 mod rules_blocks;
 mod rules_fetch;
-mod rules_markdown;
 mod rules_shapes;
 mod scoring;
