+++ skills/orch/tests/queue_wait.sh
@@ -891,6 +891,71 @@ assert_eq "$(jq -r .verdict <<<"$out")" "merged" "merged verdict unchanged by th
 assert_eq "$(jq -r 'has("progressing")' <<<"$out")" "true" "merged verdict carries progressing" "$err"
 assert_eq "$(jq -r 'has("cause")' <<<"$out")" "false" "merged verdict carries no progress cause" "$err"
 
+# Call-recording gh stub for the argument cases: every one terminates in
+# the parser, before auth or any gh call, so ANY gh invocation is a failure
+# the log makes visible.
+mkdir -p "$TMP_ROOT/argbin"
+cat > "$TMP_ROOT/argbin/gh" <<EOF
+#!/usr/bin/env bash
+printf '%s\n' "\$*" >> "$TMP_ROOT/argval-gh.calls"
+exit 1
+EOF
+chmod +x "$TMP_ROOT/argbin/gh"
+
+run_qw() {
+  (cd "$TMP_ROOT/repo" \
+    && PATH="$TMP_ROOT/argbin:$PATH" \
+       .agents/skills/orch/scripts/queue-wait "$@")
+}
+
+echo "=== queue-wait argument validation ==="
+
+# --- 15. argument validation: poll_interval > max_wait ----------
+# The reported invocation shape: `queue-wait 481 1800` reads as poll=1800,
+# max=600 and can only ever poll once while overshooting the budget.
+err="$TMP_ROOT/e15"
+run_qw 1 1800 600 --json --no-check-probe >/dev/null 2>"$err" && rc=0 || rc=$?
+assert_eq "$rc" "2" "poll_interval > max_wait exits 2" "$err"
+assert_contains "$(cat "$err")" "exceeds max_wait" "swapped-arg error names the cause" "$err"
+
+# --- 16. argument validation: non-numeric interval --------------------------
+err="$TMP_ROOT/e16"
+run_qw 1 abc 600 --json --no-check-probe >/dev/null 2>"$err" && rc=0 || rc=$?
+assert_eq "$rc" "2" "non-numeric poll_interval exits 2" "$err"
+assert_contains "$(cat "$err")" "positive integer" "non-numeric error is explicit" "$err"
+
+# --- 17. --help prints usage and exits 0 ------------------------------------
+err="$TMP_ROOT/e17"
+out="$(run_qw --help 2>"$err")" && rc=0 || rc=$?
+assert_eq "$rc" "0" "--help exits 0" "$err"
+assert_contains "$out" "Usage: queue-wait" "--help prints usage" "$err"
+# The heredoc is the contract's sole home: pin tokens whose
+# semantics live nowhere else: tokens, never sentences).
+assert_contains "$out" "Exit codes:" "--help carries the exit-code table" "$err"
+assert_contains "$out" "not_queued" "--help carries the verdict vocabulary" "$err"
+assert_contains "$out" "QUEUE_WAIT_ARM_GRACE" "--help carries the environment knobs" "$err"
+
+# --- 17b. an unknown flag is rejected in the parser, never absorbed ---------
+# Without the -*) branch, --bogus-flag lands in a positional slot and dies
+# later blaming poll_interval, same shape as ci-wait).
+err="$TMP_ROOT/e17b"
+run_qw 1 30 600 --bogus-flag >/dev/null 2>"$err" && rc=0 || rc=$?
+assert_eq "$rc" "2" "unknown flag exits 2" "$err"
+assert_contains "$(cat "$err")" "unknown option" "unknown-flag error names the flag, not a positional" "$err"
+
+
+# Missing PR# is the usage-error class (exit 2), never exit 1.
+err="$TMP_ROOT/e-missing"
+run_qw >/dev/null 2>"$err" && rc=0 || rc=$?
+assert_eq "$rc" "2" "missing PR# exits 2" "$err"
+assert_contains "$(cat "$err")" "missing required <PR#>" "missing PR# names the argument" "$err"
+
+if [[ -e "$TMP_ROOT/argval-gh.calls" ]]; then
+  assert_eq "$(cat "$TMP_ROOT/argval-gh.calls")" "" "no case above ever invoked gh"
+else
+  assert_eq "no-calls" "no-calls" "no case above ever invoked gh"
+fi
+
 echo
 printf 'pass: %d   fail: %d\n' "$PASS" "$FAIL"
 [[ "$FAIL" -eq 0 ]]
